program/crawler/requirements.txt
# 跨境电商选品系统 - 爬虫模块依赖
# 爬虫框架
scrapy==2.11.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1

# 数据处理
pandas==2.1.1
numpy==1.24.3

# 数据库连接
psycopg2-binary==2.9.7
sqlalchemy==2.0.19

# 反爬虫
fake-useragent==1.4.0
user-agents==2.2.0
rotating-proxies==0.6.2

# 数据处理和清洗
lxml==4.9.3
html5lib==1.1
pyquery==2.0.0

# 调度和任务管理
celery==5.3.1
redis==4.6.0
schedule==1.2.0

# 日志和监控
loguru==0.7.0
prometheus-client==0.17.1

# 工具库
python-dotenv==1.0.0
tqdm==4.65.0
